Aggregate data based on grouping variables and a user-provided function

Use grouping variables to collapse spectral data.frame by mean or median. Recommended for use after FilterSpectra

AggregateSpectra(df, grouping.colnames, reference.value.colname,
  agg.function)

Arguments

df
data.frame object containing one or multiple columns of grouping variables (must be consistent within each group), column of reference values (optional), and columns of spectra. Spectral column names must start with "X".

grouping.colnames
Names of columns to be used as grouping variables. Minimum 2 variables required. Default is c("trial", "plot").

reference.value.colname
Name of reference column to be aggregated along with spectra. Default is "reference"

agg.function
Name of function (string format) to be used for sample aggregation. Must be either "mean" or "median". Default is "mean".

Value

data.frame object df aggregated based on grouping column by agg.function

Author

Jenna Hershberger jmh579@cornell.edu

Examples

library(magrittr)
aggregated.test <- ikeogu.2017 %>%
  dplyr::select(-TCC) %>%
  na.omit() %>%
  AggregateSpectra(grouping.colnames = c("study.name"),
                   reference.value.colname = "DMC.oven",
                   agg.function = "mean")
aggregated.test[1:5, 1:5]

#>      study.name DMC.oven      X350      X351      X352
#> 1       C16Mcal 36.06255 0.5589216 0.5619379 0.5685708
#> 2       C16Mval 36.39604 0.5448650 0.5492415 0.5521408
#> NA         <NA>       NA        NA        NA        NA
#> NA.1       <NA>       NA        NA        NA        NA
#> NA.2       <NA>       NA        NA        NA        NA