Save spectral prediction model and model performance statistics

Saves spectral prediction model and model statistics to model.save.folder as model.name.Rds and model.name_stats.csv respectively

SaveModel(
  df,
  save.model = TRUE,
  autoselect.preprocessing = TRUE,
  preprocessing.method = NULL,
  model.save.folder = NULL,
  model.name = "PredictionModel",
  best.model.metric = "RMSE",
  tune.length = 50,
  model.method = "pls",
  num.iterations = 10,
  wavelengths = 740:1070,
  stratified.sampling = TRUE,
  cv.scheme = NULL,
  trial1 = NULL,
  trial2 = NULL,
  trial3 = NULL,
  verbose = TRUE
)

Arguments

df
data.frame object. First column contains unique identifiers, second contains reference values, followed by spectral columns. Include no other columns to right of spectra! Column names of spectra must start with "X" and reference column must be named "reference"

save.model
If TRUE, the trained model will be saved in .Rds format to the location specified by model.save.folder. If FALSE, model will be output by function but will not save to file. Default is TRUE.

autoselect.preprocessing
Boolean that, if TRUE, will choose the preprocessing method for the saved model using the best.model.metric. If FALSE, the user must supply the preprocessing method (1-12, see DoPreprocessing() documentation for more information) of the saved model. Default is TRUE.

preprocessing.method
Number or list of numbers 1:13 corresponding to desired pretreatment method(s):

1 = raw data (default)

2 = standard normal variate (SNV)

3 = SNV and first derivative

4 = SNV and second derivative

5 = first derivative

6 = second derivative

7 = Savitzky–Golay filter (SG)

8 = SNV and SG

9 = gap segment derivative (window size = 11)

10 = SG and first derivative (window size = 5)

11 = SG and first derivative (window size = 11)

12 = SG and second derivative (window size = 5)

13 = SG and second derivative (window size = 11)

model.save.folder
Path to folder where model will be saved. If not provided, will save to working directory.

model.name
Name that model will be saved as in model.save.folder. Default is "PredictionModel".

best.model.metric
Metric used to decide which model is best. Must be either "RMSE" or "Rsquared"

tune.length
Number delineating search space for tuning of the PLSR hyperparameter ncomp. Default is 50.

model.method
Model type to use for training. Valid options include:

"pls": Partial least squares regression (Default)

"rf": Random forest

"svmLinear": Support vector machine with linear kernel

"svmRadial": Support vector machine with radial kernel

num.iterations
Number of training iterations to perform

wavelengths
List of wavelengths represented by each column in df

stratified.sampling
If TRUE, training and test sets will be selected using stratified random sampling. This term is only used if test.data == NULL. Default is TRUE.

cv.scheme
A cross validation (CV) scheme from Jarquín et al., 2017. Options for cv.scheme include:

"CV1": untested lines in tested environments

"CV2": tested lines in tested environments

"CV0": tested lines in untested environments

"CV00": untested lines in untested environments

trial1
data.frame object that is for use only when cv.scheme is provided. Contains the trial to be tested in subsequent model training functions. The first column contains unique identifiers, second contains genotypes, third contains reference values, followed by spectral columns. Include no other columns to right of spectra! Column names of spectra must start with "X", reference column must be named "reference", and genotype column must be named "genotype".

trial2
data.frame object that is for use only when cv.scheme is provided. This data.frame contains a trial that has overlapping genotypes with trial1 but that were grown in a different site/year (different environment). Formatting must be consistent with trial1.

trial3
data.frame object that is for use only when cv.scheme is provided. This data.frame contains a trial that may or may not contain genotypes that overlap with trial1. Formatting must be consistent with trial1.

verbose
If TRUE, the number of rows removed through filtering will be printed to the console. Default is TRUE.

Value

List of model stats (in data.frame) and trained model object. Saves both to model.save.folder as well. To use optimally trained model for predictions, use tuned parameters from $bestTune

Details

Wrapper that uses DoPreprocessing, FormatCV, and TrainSpectralModel functions.

Author

Jenna Hershberger jmh579@cornell.edu

Examples

# \donttest{
library(magrittr)
test.model <- ikeogu.2017 %>%
  dplyr::filter(study.name == "C16Mcal") %>%
  dplyr::rename(reference = DMC.oven) %>%
  dplyr::select(sample.id, reference, dplyr::starts_with("X")) %>%
  na.omit() %>%
  SaveModel(df = ., save.model = FALSE,
            autoselect.preprocessing = TRUE,
            model.name = "my_prediction_model",
            tune.length = 50, num.iterations = 10,
            wavelengths = 350:2500)

#> Preprocessing...
#> Training model 1 of 13: (Raw_data).

#> Loading required package: lattice

#> Loading required package: ggplot2

#> 
#> Returning model...

#> Warning: provided 21 variables to replace 17 variables

#> Training model 2 of 13: (SNV).
#> 
#> Returning model...

#> Warning: provided 21 variables to replace 17 variables

#> Training model 3 of 13: (SNV1D).
#> 
#> Returning model...

#> Warning: provided 21 variables to replace 17 variables

#> Training model 4 of 13: (SNV2D).
#> 
#> Returning model...

#> Warning: provided 21 variables to replace 17 variables

#> Training model 5 of 13: (D1).
#> 
#> Returning model...

#> Warning: provided 21 variables to replace 17 variables

#> Training model 6 of 13: (D2).
#> 
#> Returning model...

#> Warning: provided 21 variables to replace 17 variables

#> Training model 7 of 13: (SG).
#> 
#> Returning model...

#> Warning: provided 21 variables to replace 17 variables

#> Training model 8 of 13: (SNVSG).
#> 
#> Returning model...

#> Warning: provided 21 variables to replace 17 variables

#> Training model 9 of 13: (SGD1).
#> 
#> Returning model...

#> Warning: provided 21 variables to replace 17 variables

#> Training model 10 of 13: (SG.D1W5).
#> 
#> Returning model...

#> Warning: provided 21 variables to replace 17 variables

#> Training model 11 of 13: (SG.D1W11).
#> 
#> Returning model...

#> Warning: provided 21 variables to replace 17 variables

#> Training model 12 of 13: (SG.D2W5).
#> 
#> Returning model...

#> Warning: provided 21 variables to replace 17 variables

#> Training model 13 of 13: (SG.D2W11).
#> 
#> Returning model...

#> Warning: provided 21 variables to replace 17 variables

#> 
#> Training Summary:
#>    Pretreatment     RMSE   Rsquared       RPD     RPIQ       CCC        Bias
#> 1      Raw_data 1.606611 0.86443860 2.6791213 3.553870 0.9188668  0.14632521
#> 2           SNV 1.617465 0.85747614 2.6517550 3.506820 0.9166633  0.08481619
#> 3         SNV1D 2.769520 0.57447384 1.5595676 2.055783 0.7273762  0.01923099
#> 4         SNV2D 4.734419 0.04473102 0.9087229 1.203707 0.1316837 -0.12981514
#> 5            D1 2.510155 0.64144653 1.7324652 2.283195 0.7704853  0.05677603
#> 6            D2 4.685534 0.05060192 0.9181383 1.216616 0.1396137 -0.06992752
#> 7            SG 1.566773 0.86871326 2.7388477 3.636040 0.9245239  0.12350017
#> 8         SNVSG 1.538586 0.86991804 2.7998941 3.709071 0.9254167  0.10337613
#> 9          SGD1 1.477693 0.88279428 2.9043920 3.874266 0.9354620  0.10573660
#> 10      SG.D1W5 2.105472 0.75857593 2.0360292 2.694982 0.8562363  0.02582945
#> 11     SG.D1W11 1.706059 0.83983349 2.5170849 3.331150 0.9102520  0.02344768
#> 12      SG.D2W5 4.431643 0.06963380 0.9698106 1.289830 0.1933924 -0.18029869
#> 13     SG.D2W11 2.916995 0.53874812 1.4771300 1.958119 0.7068506  0.02368742
#>         SEP      R2sp   RMSE.sd Rsquared.sd    RPD.sd    RPIQ.sd     CCC.sd
#> 1  1.629400 0.9664434 0.9476518  0.86757937 0.1575526 0.03475150 0.34477190
#> 2  1.640408 0.9896866 0.9447618  0.85411469 0.0862863 0.03472313 0.32957567
#> 3  2.808806 0.8121179 0.9413956  0.54684191 0.2349187 0.13583875 0.25190898
#> 4  4.801578 2.4964016 0.5976189  0.04929528 0.3968636 0.05785177 0.11922488
#> 5  2.545762 0.7490173 0.9642992  0.60622791 0.2970020 0.13547279 0.32008852
#> 6  4.751999 2.4705460 0.6040690  0.05249794 0.4040895 0.06489445 0.11811363
#> 7  1.588998 1.0117567 0.9433401  0.86373570 0.1342345 0.03303286 0.29871631
#> 8  1.560411 0.9420298 0.9514194  0.86338127 0.1461339 0.03219045 0.38360471
#> 9  1.498655 0.8614543 0.9591356  0.87171864 0.1658059 0.02143230 0.24267152
#> 10 2.135339 0.8095936 0.9576170  0.72503982 0.1132593 0.05785480 0.24911831
#> 11 1.730260 0.5190804 0.9842840  0.82103605 0.1303047 0.04260078 0.30808034
#> 12 4.494506 2.7170398 0.5332571  0.08598476 0.4683884 0.05746438 0.09927615
#> 13 2.958373 0.9513946 0.9278163  0.47037170 0.2810221 0.12799287 0.19781791
#>      Bias.sd     SEP.sd   R2sp.sd                 NA
#> 1  0.3619965 0.02553499 0.2646742  0.159787538834462
#> 2  0.1699663 0.02526888 0.2361953 0.0875102752773452
#> 3  0.1645080 0.08983745 0.5019851  0.238250999997268
#> 4  0.1112158 0.10833967 0.8178810  0.402493146506399
#> 5  0.2723527 0.09037000 0.4111572  0.301215046386371
#> 6  0.1137540 0.11576655 0.7917853  0.409821554855555
#> 7  0.3181155 0.01981273 0.2294415  0.136138676308338
#> 8  0.3620474 0.02422825 0.2512978   0.14820677727707
#> 9  0.4507682 0.01170574 0.2032037  0.168157843905494
#> 10 0.1501343 0.04148465 0.3342339  0.114865926652914
#> 11 0.2037924 0.02864509 0.2649881  0.132153066711667
#> 12 0.1367206 0.10513826 0.5955086  0.475032578600728
#> 13 0.2112801 0.09598145 0.4668611  0.285008475908858
#> 
#> Best preprocessing technique: SGD1

summary(test.model[1])

#>      Length Class      Mode
#> [1,] 18     data.frame list

test.model[2]

#> [[1]]
#> Partial least squares regression , fitted with the kernel algorithm.
#> Call:
#> plsr(formula = reference ~ spectra, ncomp = tune.length, data = df.plsr)
#> 

# }