Save spectral prediction model and model performance statistics

Given a set of pretreatment methods, saves the best spectral prediction model and model statistics to model.save.folder as model.name.Rds and model.name_stats.csv respectively. If only one pretreatment method is supplied, results from that method are stored.

save_model(
  df,
  write.model = TRUE,
  pretreatment = 1,
  model.save.folder = NULL,
  model.name = "PredictionModel",
  best.model.metric = "RMSE",
  k.folds = 5,
  proportion.train = 0.7,
  tune.length = 50,
  model.method = "pls",
  num.iterations = 10,
  stratified.sampling = TRUE,
  cv.scheme = NULL,
  trial1 = NULL,
  trial2 = NULL,
  trial3 = NULL,
  verbose = TRUE,
  save.model = deprecated(),
  wavelengths = deprecated(),
  autoselect.preprocessing = deprecated(),
  preprocessing.method = deprecated()
)

Arguments

df
data.frame object. First column contains unique identifiers, second contains reference values, followed by spectral columns. Include no other columns to right of spectra! Column names of spectra must start with "X" and reference column must be named "reference"

write.model
If TRUE, the trained model will be saved in .Rds format to the location specified by model.save.folder. If FALSE, the best model will be output by the function but will not save to a file. Default is TRUE.

pretreatment
Number or list of numbers 1:13 corresponding to desired pretreatment method(s):

Raw data (default)

Standard normal variate (SNV)

SNV and first derivative

SNV and second derivative

First derivative

Second derivative

Savitzky–Golay filter (SG)

SNV and SG

Gap segment derivative (window size = 11)

SG and first derivative (window size = 5)

SG and first derivative (window size = 11)

SG and second derivative (window size = 5)

SG and second derivative (window size = 11)

model.save.folder
Path to folder where model will be saved. If not provided, will save to working directory.

model.name
Name that model will be saved as in model.save.folder. Default is "PredictionModel".

best.model.metric
Metric used to decide which model is best. Must be either "RMSE" or "Rsquared"

k.folds
Number indicating the number of folds for k-fold cross-validation during model training. Default is 5.

proportion.train
Fraction of samples to include in the training set. Default is 0.7.

tune.length
Number delineating search space for tuning of the PLSR hyperparameter ncomp. Must be set to 5 when using the random forest algorithm (model.method == rf). Default is 50.

model.method
Model type to use for training. Valid options include:

"pls": Partial least squares regression (Default)

"rf": Random forest

"svmLinear": Support vector machine with linear kernel

"svmRadial": Support vector machine with radial kernel

num.iterations
Number of training iterations to perform

stratified.sampling
If TRUE, training and test sets will be selected using stratified random sampling. This term is only used if test.data == NULL. Default is TRUE.

cv.scheme
A cross validation (CV) scheme from Jarquín et al., 2017. Options for cv.scheme include:

"CV1": untested lines in tested environments

"CV2": tested lines in tested environments

"CV0": tested lines in untested environments

"CV00": untested lines in untested environments

trial1
data.frame object that is for use only when cv.scheme is provided. Contains the trial to be tested in subsequent model training functions. The first column contains unique identifiers, second contains genotypes, third contains reference values, followed by spectral columns. Include no other columns to right of spectra! Column names of spectra must start with "X", reference column must be named "reference", and genotype column must be named "genotype".

trial2
data.frame object that is for use only when cv.scheme is provided. This data.frame contains a trial that has overlapping genotypes with trial1 but that were grown in a different site/year (different environment). Formatting must be consistent with trial1.

trial3
data.frame object that is for use only when cv.scheme is provided. This data.frame contains a trial that may or may not contain genotypes that overlap with trial1. Formatting must be consistent with trial1.

verbose
If TRUE, the number of rows removed through filtering will be printed to the console. Default is TRUE.

save.model
`r lifecycle::badge("deprecated")` save.model = FALSE is no longer supported; this function will always return a saved model.

wavelengths
`r lifecycle::badge("deprecated")` wavelengths is no longer supported; this information is now inferred from df column names

autoselect.preprocessing
`r lifecycle::badge("deprecated")` autoselect.preprocessing = FALSE is no longer supported. If multiple pretreatment methods are supplied, the best will be automatically selected as the model to be saved.

preprocessing.method
`r lifecycle::badge("deprecated")` this argument has been renamed "pretreatment"

Value

List of model stats (in data.frame) and trained model object. If the parameter write.model is TRUE, both objects are saved to model.save.folder. To use the optimally trained model for predictions, use tuned parameters from $bestTune.

Details

Wrapper that uses pretreat_spectra, format_cv, and train_spectra functions.

Author

Jenna Hershberger jmh579@cornell.edu

Examples

# \donttest{
library(magrittr)
test.model <- ikeogu.2017 %>%
  dplyr::filter(study.name == "C16Mcal") %>%
  dplyr::rename(reference = DMC.oven,
                unique.id = sample.id) %>%
  dplyr::select(unique.id, reference, dplyr::starts_with("X")) %>%
  na.omit() %>%
  save_model(
    df = .,
    write.model = FALSE,
    pretreatment = 1:13,
    model.name = "my_prediction_model",
    tune.length = 50,
    num.iterations = 10
  )

#> Pretreatment initiated.
#> Training models...
#> Working on Raw_data 

#> Loading required package: lattice

#> 
#> Attaching package: ‘pls’

#> The following object is masked from ‘package:caret’:
#> 
#>     R2

#> The following object is masked from ‘package:stats’:
#> 
#>     loadings

#> Returning model...
#> Working on SNV 
#> Returning model...
#> Working on SNV1D 
#> Returning model...
#> Working on SNV2D 
#> Returning model...
#> Working on D1 
#> Returning model...
#> Working on D2 
#> Returning model...
#> Working on SG 
#> Returning model...
#> Working on SNVSG 
#> Returning model...
#> Working on SGD1 
#> Returning model...
#> Working on SG.D1W5 
#> Returning model...
#> Working on SG.D1W11 
#> Returning model...
#> Working on SG.D2W5 
#> Returning model...
#> Working on SG.D2W11 
#> Returning model...
#> 
#> Training Summary:
#> # A tibble: 13 x 40
#>    Pretreatment RMSEp.mean R2p.mean RPD.mean RPIQ.mean CCC.mean Bias.mean
#>    <fct>             <dbl>    <dbl>    <dbl>     <dbl>    <dbl>     <dbl>
#>  1 Raw_data           1.45   0.888     2.95       3.88    0.938 -0.0641  
#>  2 SNV                1.43   0.887     3.02       3.86    0.940 -0.102   
#>  3 SNV1D              2.65   0.637     1.64       2.12    0.773  0.123   
#>  4 SNV2D              4.46   0.0427    0.966      1.25    0.154  0.159   
#>  5 D1                 2.48   0.668     1.76       2.26    0.797  0.121   
#>  6 D2                 4.41   0.0582    0.976      1.26    0.182  0.209   
#>  7 SG                 1.39   0.894     3.13       4.00    0.942 -0.0545  
#>  8 SNVSG              1.44   0.884     3.02       3.86    0.937 -0.0387  
#>  9 SGD1               1.45   0.886     2.97       3.81    0.936 -0.000519
#> 10 SG.D1W5            2.00   0.787     2.18       2.79    0.879  0.0226  
#> 11 SG.D1W11           1.69   0.842     2.57       3.29    0.913 -0.103   
#> 12 SG.D2W5            4.05   0.140     1.07       1.38    0.281  0.0229  
#> 13 SG.D2W11           2.77   0.610     1.58       2.03    0.753  0.354   
#> # … with 33 more variables: SEP.mean <dbl>, RMSEcv.mean <dbl>, R2cv.mean <dbl>,
#> #   R2sp.mean <dbl>, best.ncomp.mean <dbl>, best.ntree.mean <dbl>,
#> #   best.mtry.mean <dbl>, RMSEp.sd <dbl>, R2p.sd <dbl>, RPD.sd <dbl>,
#> #   RPIQ.sd <dbl>, CCC.sd <dbl>, Bias.sd <dbl>, SEP.sd <dbl>, RMSEcv.sd <dbl>,
#> #   R2cv.sd <dbl>, R2sp.sd <dbl>, best.ncomp.sd <dbl>, best.ntree.sd <dbl>,
#> #   best.mtry.sd <dbl>, RMSEp.mode <dbl>, R2p.mode <dbl>, RPD.mode <dbl>,
#> #   RPIQ.mode <dbl>, CCC.mode <dbl>, Bias.mode <dbl>, SEP.mode <dbl>,
#> #   RMSEcv.mode <dbl>, R2cv.mode <dbl>, R2sp.mode <dbl>, best.ncomp.mode <dbl>,
#> #   best.ntree.mode <dbl>, best.mtry.mode <dbl>
#> 
#> Best pretreatment technique: SG

summary(test.model$best.model)

#> Data: 	X dimension: 120 2141 
#> 	Y dimension: 120 1
#> Fit method: kernelpls
#> Number of components considered: 50
#> TRAINING: % variance explained
#>            1 comps  2 comps  3 comps  4 comps  5 comps  6 comps  7 comps
#> X            62.42    72.33    94.08    97.08    97.59    98.56    99.14
#> reference    66.31    75.59    77.69    81.23    88.74    89.96    90.84
#>            8 comps  9 comps  10 comps  11 comps  12 comps  13 comps  14 comps
#> X            99.54    99.79     99.86     99.87     99.90     99.93     99.93
#> reference    91.48    92.40     92.99     93.93     94.29     94.64     95.70
#>            15 comps  16 comps  17 comps  18 comps  19 comps  20 comps  21 comps
#> X             99.94     99.95     99.96     99.97     99.97     99.98     99.98
#> reference     96.08     96.37     96.68     97.22     97.68     97.98     98.34
#>            22 comps  23 comps  24 comps  25 comps  26 comps  27 comps  28 comps
#> X             99.98     99.98     99.98     99.98     99.99     99.99     99.99
#> reference     98.62     98.81     98.96     99.13     99.22     99.40     99.51
#>            29 comps  30 comps  31 comps  32 comps  33 comps  34 comps  35 comps
#> X             99.99     99.99     99.99     99.99     99.99     99.99     99.99
#> reference     99.59     99.67     99.74     99.80     99.85     99.89     99.93
#>            36 comps  37 comps  38 comps  39 comps  40 comps  41 comps  42 comps
#> X             99.99     99.99     99.99     99.99     99.99     99.99     99.99
#> reference     99.94     99.95     99.97     99.98     99.98     99.99     99.99
#>            43 comps  44 comps  45 comps  46 comps  47 comps  48 comps  49 comps
#> X             99.99     99.99     99.99     99.99       100       100       100
#> reference     99.99    100.00    100.00    100.00       100       100       100
#>            50 comps
#> X               100
#> reference       100

test.model$best.model.stats

#> # A tibble: 1 x 40
#>   Pretreatment RMSEp.mean R2p.mean RPD.mean RPIQ.mean CCC.mean Bias.mean
#>   <fct>             <dbl>    <dbl>    <dbl>     <dbl>    <dbl>     <dbl>
#> 1 SG                 1.39    0.894     3.13      4.00    0.942   -0.0545
#> # … with 33 more variables: SEP.mean <dbl>, RMSEcv.mean <dbl>, R2cv.mean <dbl>,
#> #   R2sp.mean <dbl>, best.ncomp.mean <dbl>, best.ntree.mean <dbl>,
#> #   best.mtry.mean <dbl>, RMSEp.sd <dbl>, R2p.sd <dbl>, RPD.sd <dbl>,
#> #   RPIQ.sd <dbl>, CCC.sd <dbl>, Bias.sd <dbl>, SEP.sd <dbl>, RMSEcv.sd <dbl>,
#> #   R2cv.sd <dbl>, R2sp.sd <dbl>, best.ncomp.sd <dbl>, best.ntree.sd <dbl>,
#> #   best.mtry.sd <dbl>, RMSEp.mode <dbl>, R2p.mode <dbl>, RPD.mode <dbl>,
#> #   RPIQ.mode <dbl>, CCC.mode <dbl>, Bias.mode <dbl>, SEP.mode <dbl>,
#> #   RMSEcv.mode <dbl>, R2cv.mode <dbl>, R2sp.mode <dbl>, best.ncomp.mode <dbl>,
#> #   best.ntree.mode <dbl>, best.mtry.mode <dbl>

# }