Test the performance of spectral models

Wrapper that trains models based spectral data to predict reference values and reports model performance statistics

test_spectra(
  train.data,
  num.iterations,
  test.data = NULL,
  pretreatment = 1,
  k.folds = 5,
  proportion.train = 0.7,
  tune.length = 50,
  model.method = "pls",
  best.model.metric = "RMSE",
  stratified.sampling = TRUE,
  cv.scheme = NULL,
  trial1 = NULL,
  trial2 = NULL,
  trial3 = NULL,
  split.test = FALSE,
  verbose = TRUE,
  wavelengths = deprecated(),
  preprocessing = deprecated(),
  output.summary = deprecated(),
  rf.variable.importance = deprecated()
)

Arguments

train.data

data.frame object of spectral data for input into a spectral prediction model. First column contains unique identifiers, second contains reference values, followed by spectral columns. Include no other columns to right of spectra! Column names of spectra must start with "X" and reference column must be named "reference".

num.iterations

Number of training iterations to perform

test.data

data.frame with same specifications as df. Use if specific test set is desired for hyperparameter tuning. If NULL, function will automatically train with a stratified sample of 70%. Default is NULL.

pretreatment

Number or list of numbers 1:13 corresponding to desired pretreatment method(s):

Raw data (default)

Standard normal variate (SNV)

SNV and first derivative

SNV and second derivative

First derivative

Second derivative

Savitzky–Golay filter (SG)

SNV and SG

Gap-segment derivative (window size = 11)

SG and first derivative (window size = 5)

SG and first derivative (window size = 11)

SG and second derivative (window size = 5)

SG and second derivative (window size = 11)

k.folds

Number indicating the number of folds for k-fold cross-validation during model training. Default is 5.

proportion.train

Fraction of samples to include in the training set. Default is 0.7.

tune.length

Number delineating search space for tuning of the PLSR hyperparameter ncomp. Must be set to 5 when using the random forest algorithm (model.method == rf). Default is 50.

model.method

Model type to use for training. Valid options include:

"pls": Partial least squares regression (Default)

"rf": Random forest

"svmLinear": Support vector machine with linear kernel

"svmRadial": Support vector machine with radial kernel

best.model.metric

Metric used to decide which model is best. Must be either "RMSE" or "Rsquared"

stratified.sampling

If TRUE, training and test sets will be selected using stratified random sampling. This term is only used if test.data == NULL. Default is TRUE.

cv.scheme

A cross validation (CV) scheme from Jarquín et al., 2017. Options for cv.scheme include:

"CV1": untested lines in tested environments

"CV2": tested lines in tested environments

"CV0": tested lines in untested environments

"CV00": untested lines in untested environments

trial1

data.frame object that is for use only when cv.scheme is provided. Contains the trial to be tested in subsequent model training functions. The first column contains unique identifiers, second contains genotypes, third contains reference values, followed by spectral columns. Include no other columns to right of spectra! Column names of spectra must start with "X", reference column must be named "reference", and genotype column must be named "genotype".

trial2

data.frame object that is for use only when cv.scheme is provided. This data.frame contains a trial that has overlapping genotypes with trial1 but that were grown in a different site/year (different environment). Formatting must be consistent with trial1.

trial3

data.frame object that is for use only when cv.scheme is provided. This data.frame contains a trial that may or may not contain genotypes that overlap with trial1. Formatting must be consistent with trial1.

split.test

boolean that allows for a fixed training set and a split test set. Example// train model on data from two breeding programs and a stratified subset (70%) of a third and test on the remaining samples (30%) of the third. If FALSE, the entire provided test set test.data will remain as a testing set or if none is provided, 30% of the provided train.data will be used for testing. Default is FALSE.

verbose

If TRUE, the number of rows removed through filtering will be printed to the console. Default is TRUE.

wavelengths

`r lifecycle::badge("deprecated")` wavelengths is no longer supported; this information is now inferred from df column names

preprocessing

`r lifecycle::badge("deprecated")` please use pretreatment to specify the specific pretreatment(s) to test. For behavior identical to that of preprocessing = TRUE, set pretreatment = 1:13`.

output.summary

`r lifecycle::badge("deprecated")` output.summary = FALSE is no longer supported; a summary of output is always returned alongside the full performance statistics.

rf.variable.importance

`r lifecycle::badge("deprecated")` rf.variable.importance = FALSE is no longer supported; variable importance results are always returned if the model.method is set to `pls` or `rf`.

Value

list of 5 objects:

`model.list` is a list of trained model objects, one for each pretreatment method specified by the pretreatment argument. Each model is trained with all rows of df.

`summary.model.performance` is a data.frame containing summary statistics across all model training iterations and pretreatments. See below for a description of the summary statistics provided.

`model.performance` is a data.frame containing performance statistics for each iteration of model training separately (see below).

`predictions` is a data.frame containing both reference and predicted values for each test set entry in each iteration of model training.

`importance` is a data.frame containing variable importance results for each wavelength at each iteration of model training. If model.method is not "pls" or "rf", this list item is NULL.

`summary.model.performance` and `model.performance` data.frames summary statistics include:

Tuned parameters depending on the model algorithm:

Best.n.comp, the best number of components

Best.ntree, the best number of trees in an RF model

Best.mtry, the best number of variables to include at every decision point in an RF model

RMSECV, the root mean squared error of cross-validation

R2cv, the coefficient of multiple determination of cross-validation for PLSR models

RMSEP, the root mean squared error of prediction

R2p, the squared Pearson’s correlation between predicted and observed test set values

RPD, the ratio of standard deviation of observed test set values to RMSEP

RPIQ, the ratio of performance to interquartile difference

CCC, the concordance correlation coefficient

Bias, the average difference between the predicted and observed values

SEP, the standard error of prediction

R2sp, the squared Spearman’s rank correlation between predicted and observed test set values

Details

Calls pretreat_spectra, format_cv, and train_spectra functions.

Author

Jenna Hershberger jmh579@cornell.edu

Examples

# \donttest{
library(magrittr)
ikeogu.2017 %>%
  dplyr::rename(reference = DMC.oven,
                unique.id = sample.id) %>%
  dplyr::select(unique.id, reference, dplyr::starts_with("X")) %>%
  na.omit() %>%
  test_spectra(
    train.data = .,
    tune.length = 3,
    num.iterations = 3,
    pretreatment = 1
  )
#> Pretreatment initiated.
#> Training models...
#> Working on Raw_data 
#> Returning model...
#> $model
#> Partial least squares regression , fitted with the kernel algorithm.
#> Call:
#> plsr(formula = reference ~ spectra, ncomp = tune.length, data = df.plsr)
#> 
#> $summary.model.performance
#>   SummaryType ModelType     RMSEp       R2p      RPD      RPIQ        CCC
#> 1        mean       pls 1.9340776 0.8045359 2.325700 2.7580330 0.88336248
#> 2          sd       pls 0.3412865 0.0649400 0.485082 0.4322122 0.04486757
#> 3        mode       pls 2.2791075 0.7552351 2.023697 2.2910671 0.86571303
#>          Bias       SEP     RMSEcv       R2cv       R2sp best.ncomp best.ntree
#> 1  0.07671425 1.9541204 2.01072530 0.75374306 0.78377823          3         NA
#> 2  0.22946500 0.3448233 0.09669301 0.02328558 0.04515009          0         NA
#> 3 -0.17201453 2.3027258 1.89939847 0.77126178 0.78123953          3         NA
#>   best.mtry
#> 1        NA
#> 2        NA
#> 3        NA
#> 
#> $model.performance
#>   Iteration ModelType    RMSEp       R2p      RPD     RPIQ       CCC       Bias
#> 1         1       pls 2.279107 0.7552351 2.023697 2.291067 0.8657130 -0.1720145
#> 2         2       pls 1.596662 0.8781176 2.885236 3.144043 0.9343710  0.1219921
#> 3         3       pls 1.926463 0.7802548 2.068167 2.838989 0.8500034  0.2801652
#>        SEP   RMSEcv      R2cv      R2sp best.ncomp best.ntree best.mtry
#> 1 2.302726 1.899398 0.7712618 0.7812395          3         NA        NA
#> 2 1.613208 2.073757 0.7273191 0.8301441          3         NA        NA
#> 3 1.946427 2.059020 0.7626483 0.7399511          3         NA        NA
#> 
#> $predictions
#>        Iteration ModelType unique.id     reference    predicted         
#>   [1,] "1"       "pls"     "C16Mcal_4"   "39.0099869" "36.6186175872549"
#>   [2,] "1"       "pls"     "C16Mcal_8"   "26.2774601" "29.870694142597" 
#>   [3,] "1"       "pls"     "C16Mcal_10"  "31.7993336" "33.8749877349711"
#>   [4,] "1"       "pls"     "C16Mcal_16"  "36.3796349" "38.4563811201679"
#>   [5,] "1"       "pls"     "C16Mcal_20"  "34.1300011" "31.7763257849577"
#>   [6,] "1"       "pls"     "C16Mcal_22"  "41.2799988" "40.2829451936288"
#>   [7,] "1"       "pls"     "C16Mcal_23"  "37.1399994" "36.7590846026353"
#>   [8,] "1"       "pls"     "C16Mcal_32"  "37.4700012" "38.0274443718859"
#>   [9,] "1"       "pls"     "C16Mcal_33"  "34.9757233" "33.3547131769862"
#>  [10,] "1"       "pls"     "C16Mcal_34"  "42.5299988" "38.9034763429464"
#>  [11,] "1"       "pls"     "C16Mcal_37"  "36.7437744" "36.9457670330436"
#>  [12,] "1"       "pls"     "C16Mcal_38"  "33.7883987" "36.0054342713846"
#>  [13,] "1"       "pls"     "C16Mcal_41"  "39.9312401" "39.3631194829371"
#>  [14,] "1"       "pls"     "C16Mcal_49"  "38.2799988" "37.7255187233156"
#>  [15,] "1"       "pls"     "C16Mcal_50"  "37.7099991" "35.196003059161" 
#>  [16,] "1"       "pls"     "C16Mcal_54"  "35.0544205" "34.6621960721599"
#>  [17,] "1"       "pls"     "C16Mcal_59"  "28.9799995" "36.0873318537158"
#>  [18,] "1"       "pls"     "C16Mcal_62"  "29.073864"  "25.5468406533992"
#>  [19,] "1"       "pls"     "C16Mcal_66"  "41.7697029" "38.6588854053873"
#>  [20,] "1"       "pls"     "C16Mcal_68"  "33.1699982" "33.6216317989734"
#>  [21,] "1"       "pls"     "C16Mcal_70"  "39.5100403" "37.9863344466999"
#>  [22,] "1"       "pls"     "C16Mcal_81"  "37.6691475" "38.8170330545381"
#>  [23,] "1"       "pls"     "C16Mcal_85"  "33.1330147" "32.552383095351" 
#>  [24,] "1"       "pls"     "C16Mcal_87"  "35.0520935" "37.967893706147" 
#>  [25,] "1"       "pls"     "C16Mcal_88"  "34.6599998" "36.1760996492566"
#>  [26,] "1"       "pls"     "C16Mcal_89"  "40.6085129" "39.0668382752329"
#>  [27,] "1"       "pls"     "C16Mcal_93"  "34.8271332" "35.8797826842083"
#>  [28,] "1"       "pls"     "C16Mcal_96"  "36.2366486" "39.3223227074157"
#>  [29,] "1"       "pls"     "C16Mcal_101" "41.6001511" "39.5104228617055"
#>  [30,] "1"       "pls"     "C16Mcal_105" "41.1385422" "37.7734531696162"
#>  [31,] "1"       "pls"     "C16Mcal_107" "38.8699989" "37.5000986674449"
#>  [32,] "1"       "pls"     "C16Mcal_110" "23.5921307" "18.957745129433" 
#>  [33,] "1"       "pls"     "C16Mcal_119" "31.1599998" "28.5720406570201"
#>  [34,] "1"       "pls"     "C16Mval_1"   "44.1297035" "41.1714361827793"
#>  [35,] "1"       "pls"     "C16Mval_3"   "43.3628883" "41.9204866306733"
#>  [36,] "1"       "pls"     "C16Mval_4"   "43.1399994" "40.5606794664868"
#>  [37,] "1"       "pls"     "C16Mval_7"   "40.4313164" "39.6483549233926"
#>  [38,] "1"       "pls"     "C16Mval_11"  "38.898819"  "38.1761830334952"
#>  [39,] "1"       "pls"     "C16Mval_12"  "38.6911316" "39.5889150428285"
#>  [40,] "1"       "pls"     "C16Mval_16"  "38.4863548" "38.3776659638643"
#>  [41,] "1"       "pls"     "C16Mval_18"  "38.1199989" "39.4858910247899"
#>  [42,] "1"       "pls"     "C16Mval_29"  "36.2344589" "36.4252063182386"
#>  [43,] "1"       "pls"     "C16Mval_35"  "35.150135"  "35.3983988864924"
#>  [44,] "1"       "pls"     "C16Mval_37"  "34.9873123" "36.2819492106078"
#>  [45,] "1"       "pls"     "C16Mval_40"  "34.2991219" "34.3485023073428"
#>  [46,] "1"       "pls"     "C16Mval_44"  "33.7523422" "34.736180139801" 
#>  [47,] "1"       "pls"     "C16Mval_47"  "31.1025772" "34.4888812187884"
#>  [48,] "1"       "pls"     "C16Mval_50"  "29.9972706" "33.2390545573886"
#>  [49,] "1"       "pls"     "C16Mval_53"  "27.3490391" "27.587073394072" 
#>  [50,] "2"       "pls"     "C16Mcal_12"  "41.9791336" "40.8072945166818"
#>  [51,] "2"       "pls"     "C16Mcal_15"  "31.6499996" "32.2909459608052"
#>  [52,] "2"       "pls"     "C16Mcal_18"  "39.5632248" "37.808264088059" 
#>  [53,] "2"       "pls"     "C16Mcal_23"  "37.1399994" "36.7511900734977"
#>  [54,] "2"       "pls"     "C16Mcal_24"  "42.1911201" "39.5387192104054"
#>  [55,] "2"       "pls"     "C16Mcal_30"  "30.9300003" "30.284776088641" 
#>  [56,] "2"       "pls"     "C16Mcal_33"  "34.9757233" "33.6239780928958"
#>  [57,] "2"       "pls"     "C16Mcal_40"  "36.1300011" "37.1672404270967"
#>  [58,] "2"       "pls"     "C16Mcal_42"  "34.7200012" "33.888887738045" 
#>  [59,] "2"       "pls"     "C16Mcal_43"  "37.9103165" "39.2140209583749"
#>  [60,] "2"       "pls"     "C16Mcal_46"  "32.3829803" "33.5985277566305"
#>  [61,] "2"       "pls"     "C16Mcal_55"  "28.1499996" "31.595504411921" 
#>  [62,] "2"       "pls"     "C16Mcal_61"  "39.4031334" "40.0380294051205"
#>  [63,] "2"       "pls"     "C16Mcal_64"  "27.2600002" "28.2768682606361"
#>  [64,] "2"       "pls"     "C16Mcal_65"  "38.8069687" "37.1167235500758"
#>  [65,] "2"       "pls"     "C16Mcal_66"  "41.7697029" "39.1013809053809"
#>  [66,] "2"       "pls"     "C16Mcal_73"  "29.9218559" "31.2403925454157"
#>  [67,] "2"       "pls"     "C16Mcal_75"  "37.4979515" "39.8607244154279"
#>  [68,] "2"       "pls"     "C16Mcal_81"  "37.6691475" "38.5594626592558"
#>  [69,] "2"       "pls"     "C16Mcal_84"  "34.1899986" "34.4795394897491"
#>  [70,] "2"       "pls"     "C16Mcal_85"  "33.1330147" "32.9536828497842"
#>  [71,] "2"       "pls"     "C16Mcal_88"  "34.6599998" "36.5331780622494"
#>  [72,] "2"       "pls"     "C16Mcal_91"  "37.9599991" "38.1516856617191"
#>  [73,] "2"       "pls"     "C16Mcal_96"  "36.2366486" "39.0333125990401"
#>  [74,] "2"       "pls"     "C16Mcal_98"  "38.8300018" "40.2304863691086"
#>  [75,] "2"       "pls"     "C16Mcal_101" "41.6001511" "39.8955317196879"
#>  [76,] "2"       "pls"     "C16Mcal_103" "39.3523369" "39.0919790136024"
#>  [77,] "2"       "pls"     "C16Mcal_104" "35.1094627" "34.3178024579414"
#>  [78,] "2"       "pls"     "C16Mcal_112" "40.9370956" "38.9502720445222"
#>  [79,] "2"       "pls"     "C16Mcal_116" "43.2331314" "41.9422975906542"
#>  [80,] "2"       "pls"     "C16Mcal_117" "40.3080254" "39.309844466162" 
#>  [81,] "2"       "pls"     "C16Mcal_118" "39.9399986" "40.2422270321641"
#>  [82,] "2"       "pls"     "C16Mcal_120" "20.1389332" "17.5584206945611"
#>  [83,] "2"       "pls"     "C16Mval_6"   "41.4940834" "39.2445893519563"
#>  [84,] "2"       "pls"     "C16Mval_11"  "38.898819"  "38.1765529079855"
#>  [85,] "2"       "pls"     "C16Mval_13"  "38.6752853" "37.7578346009011"
#>  [86,] "2"       "pls"     "C16Mval_18"  "38.1199989" "39.8097285542192"
#>  [87,] "2"       "pls"     "C16Mval_20"  "37.9215317" "37.6467759543947"
#>  [88,] "2"       "pls"     "C16Mval_26"  "36.4902077" "37.1500127315042"
#>  [89,] "2"       "pls"     "C16Mval_27"  "36.4665642" "38.9591000285892"
#>  [90,] "2"       "pls"     "C16Mval_28"  "36.3800011" "36.9500343248765"
#>  [91,] "2"       "pls"     "C16Mval_31"  "36.0321198" "35.2065087414001"
#>  [92,] "2"       "pls"     "C16Mval_33"  "35.4645767" "33.7648517045387"
#>  [93,] "2"       "pls"     "C16Mval_35"  "35.150135"  "35.4292710135701"
#>  [94,] "2"       "pls"     "C16Mval_42"  "33.8788452" "35.5097898846799"
#>  [95,] "2"       "pls"     "C16Mval_44"  "33.7523422" "34.8615783932716"
#>  [96,] "2"       "pls"     "C16Mval_49"  "30.8113632" "33.1631336938011"
#>  [97,] "2"       "pls"     "C16Mval_50"  "29.9972706" "33.7888760480644"
#>  [98,] "2"       "pls"     "C16Mval_53"  "27.3490391" "27.668024256398" 
#>  [99,] "3"       "pls"     "C16Mcal_1"   "39.6210938" "37.59831525137"  
#> [100,] "3"       "pls"     "C16Mcal_7"   "35.8100014" "37.797232070859" 
#> [101,] "3"       "pls"     "C16Mcal_8"   "26.2774601" "30.2148223992849"
#> [102,] "3"       "pls"     "C16Mcal_25"  "31.7656345" "33.6320488966618"
#> [103,] "3"       "pls"     "C16Mcal_28"  "29.2099991" "31.8708056628753"
#> [104,] "3"       "pls"     "C16Mcal_30"  "30.9300003" "30.9745512126418"
#> [105,] "3"       "pls"     "C16Mcal_32"  "37.4700012" "38.1881386999919"
#> [106,] "3"       "pls"     "C16Mcal_37"  "36.7437744" "36.9177881856886"
#> [107,] "3"       "pls"     "C16Mcal_42"  "34.7200012" "33.9809641335401"
#> [108,] "3"       "pls"     "C16Mcal_43"  "37.9103165" "39.3242632210265"
#> [109,] "3"       "pls"     "C16Mcal_45"  "29.9400005" "32.7292422067887"
#> [110,] "3"       "pls"     "C16Mcal_49"  "38.2799988" "37.8872548336818"
#> [111,] "3"       "pls"     "C16Mcal_51"  "39.348053"  "37.4398265877267"
#> [112,] "3"       "pls"     "C16Mcal_54"  "35.0544205" "34.6820521141834"
#> [113,] "3"       "pls"     "C16Mcal_56"  "38.4799995" "38.1495885281911"
#> [114,] "3"       "pls"     "C16Mcal_57"  "40.5781174" "38.1378737729733"
#> [115,] "3"       "pls"     "C16Mcal_58"  "39.8833466" "39.5751942330403"
#> [116,] "3"       "pls"     "C16Mcal_60"  "36.6251221" "37.0864874695333"
#> [117,] "3"       "pls"     "C16Mcal_62"  "29.073864"  "26.2129904429798"
#> [118,] "3"       "pls"     "C16Mcal_65"  "38.8069687" "37.2405401176339"
#> [119,] "3"       "pls"     "C16Mcal_76"  "35.254776"  "34.577918985565" 
#> [120,] "3"       "pls"     "C16Mcal_78"  "39.3699989" "38.7776280558322"
#> [121,] "3"       "pls"     "C16Mcal_79"  "35.9900017" "39.3265967142223"
#> [122,] "3"       "pls"     "C16Mcal_86"  "39.5099983" "41.4283315483407"
#> [123,] "3"       "pls"     "C16Mcal_89"  "40.6085129" "38.9369763989427"
#> [124,] "3"       "pls"     "C16Mcal_93"  "34.8271332" "36.006660628509" 
#> [125,] "3"       "pls"     "C16Mcal_96"  "36.2366486" "39.1486970669202"
#> [126,] "3"       "pls"     "C16Mcal_97"  "34.2159386" "35.3133099288265"
#> [127,] "3"       "pls"     "C16Mcal_105" "41.1385422" "37.768790989719" 
#> [128,] "3"       "pls"     "C16Mcal_106" "40.3611107" "37.0304969698495"
#> [129,] "3"       "pls"     "C16Mcal_108" "39.1699982" "38.416057279318" 
#> [130,] "3"       "pls"     "C16Mcal_109" "28.9400005" "32.1996382710298"
#> [131,] "3"       "pls"     "C16Mcal_111" "37.6399994" "36.6313854655834"
#> [132,] "3"       "pls"     "C16Mcal_117" "40.3080254" "39.1147197347587"
#> [133,] "3"       "pls"     "C16Mval_2"   "43.7411308" "40.66869967245"  
#> [134,] "3"       "pls"     "C16Mval_4"   "43.1399994" "40.6118392939509"
#> [135,] "3"       "pls"     "C16Mval_5"   "41.7549706" "41.4848158537502"
#> [136,] "3"       "pls"     "C16Mval_18"  "38.1199989" "39.6179915506997"
#> [137,] "3"       "pls"     "C16Mval_21"  "37.8402519" "38.0890369826939"
#> [138,] "3"       "pls"     "C16Mval_31"  "36.0321198" "35.218487517254" 
#> [139,] "3"       "pls"     "C16Mval_35"  "35.150135"  "35.504278758319" 
#> [140,] "3"       "pls"     "C16Mval_37"  "34.9873123" "36.3014212037293"
#> [141,] "3"       "pls"     "C16Mval_41"  "34.1728096" "35.3830690966915"
#> [142,] "3"       "pls"     "C16Mval_42"  "33.8788452" "35.6759886108811"
#> [143,] "3"       "pls"     "C16Mval_43"  "33.8304443" "34.2946565516045"
#> [144,] "3"       "pls"     "C16Mval_44"  "33.7523422" "34.8497749841793"
#> [145,] "3"       "pls"     "C16Mval_45"  "33.442482"  "35.6549241124401"
#> [146,] "3"       "pls"     "C16Mval_47"  "31.1025772" "34.6640279616496"
#> [147,] "3"       "pls"     "C16Mval_49"  "30.8113632" "33.2475345506149"
#> 
#> $importance
#> # A tibble: 3 × 2,153
#>   Iteration ModelType   X350   X351   X352   X353   X354   X355   X356   X357
#>       <int> <chr>      <dbl>  <dbl>  <dbl>  <dbl>  <dbl>  <dbl>  <dbl>  <dbl>
#> 1         1 pls       0.0281 0.0293 0.0338 0.0319 0.0292 0.0329 0.0359 0.0340
#> 2         2 pls       0.0355 0.0355 0.0338 0.0324 0.0343 0.0390 0.0442 0.0398
#> 3         3 pls       0.0274 0.0279 0.0272 0.0265 0.0274 0.0314 0.0340 0.0329
#> # … with 2,143 more variables: X358 <dbl>, X359 <dbl>, X360 <dbl>, X361 <dbl>,
#> #   X362 <dbl>, X363 <dbl>, X364 <dbl>, X365 <dbl>, X366 <dbl>, X367 <dbl>,
#> #   X368 <dbl>, X369 <dbl>, X370 <dbl>, X371 <dbl>, X372 <dbl>, X373 <dbl>,
#> #   X374 <dbl>, X375 <dbl>, X376 <dbl>, X377 <dbl>, X378 <dbl>, X379 <dbl>,
#> #   X380 <dbl>, X381 <dbl>, X382 <dbl>, X383 <dbl>, X384 <dbl>, X385 <dbl>,
#> #   X386 <dbl>, X387 <dbl>, X388 <dbl>, X389 <dbl>, X390 <dbl>, X391 <dbl>,
#> #   X392 <dbl>, X393 <dbl>, X394 <dbl>, X395 <dbl>, X396 <dbl>, X397 <dbl>, …
#> 
# }