Train a model based predict reference values with spectral data

Trains spectral prediction models using one of several algorithms and sampling procedures.

Get the mode of a set of numbers. Used in getting summary of results within [train_spectra()]

train_spectra(
  df,
  num.iterations,
  test.data = NULL,
  k.folds = 5,
  proportion.train = 0.7,
  tune.length = 50,
  model.method = "pls",
  best.model.metric = "RMSE",
  stratified.sampling = TRUE,
  cv.scheme = NULL,
  trial1 = NULL,
  trial2 = NULL,
  trial3 = NULL,
  split.test = FALSE,
  seed = 1,
  verbose = TRUE,
  rf.variable.importance = deprecated(),
  output.summary = deprecated(),
  return.model = deprecated()
)

Arguments

df

data.frame object. First column contains unique identifiers, second contains reference values, followed by spectral columns. Include no other columns to right of spectra! Column names of spectra must start with "X" and reference column must be named "reference"

num.iterations

Number of training iterations to perform

test.data

data.frame with same specifications as df. Use if specific test set is desired for hyperparameter tuning. If NULL, function will automatically train with a stratified sample of 70%. Default is NULL.

k.folds

Number indicating the number of folds for k-fold cross-validation during model training. Default is 5.

proportion.train

Fraction of samples to include in the training set. Default is 0.7.

tune.length

Number delineating search space for tuning of the PLSR hyperparameter ncomp. Must be set to 5 when using the random forest algorithm (model.method == rf). Default is 50.

model.method

Model type to use for training. Valid options include:

"pls": Partial least squares regression (Default)

"rf": Random forest

"svmLinear": Support vector machine with linear kernel

"svmRadial": Support vector machine with radial kernel

best.model.metric

Metric used to decide which model is best. Must be either "RMSE" or "Rsquared"

stratified.sampling

If TRUE, training and test sets will be selected using stratified random sampling. This term is only used if test.data == NULL. Default is TRUE.

cv.scheme

A cross validation (CV) scheme from Jarquín et al., 2017. Options for cv.scheme include:

"CV1": untested lines in tested environments

"CV2": tested lines in tested environments

"CV0": tested lines in untested environments

"CV00": untested lines in untested environments

trial1

data.frame object that is for use only when cv.scheme is provided. Contains the trial to be tested in subsequent model training functions. The first column contains unique identifiers, second contains genotypes, third contains reference values, followed by spectral columns. Include no other columns to right of spectra! Column names of spectra must start with "X", reference column must be named "reference", and genotype column must be named "genotype".

trial2

data.frame object that is for use only when cv.scheme is provided. This data.frame contains a trial that has overlapping genotypes with trial1 but that were grown in a different site/year (different environment). Formatting must be consistent with trial1.

trial3

data.frame object that is for use only when cv.scheme is provided. This data.frame contains a trial that may or may not contain genotypes that overlap with trial1. Formatting must be consistent with trial1.

split.test

boolean that allows for a fixed training set and a split test set. Example// train model on data from two breeding programs and a stratified subset (70%) of a third and test on the remaining samples (30%) of the third. If FALSE, the entire provided test set test.data will remain as a testing set or if none is provided, 30% of the provided train.data will be used for testing. Default is FALSE.

seed

Integer to be used internally as input for set.seed(). Only used if stratified.sampling = TRUE. In all other cases, seed is set to the current iteration number. Default is 1.

verbose

If TRUE, the number of rows removed through filtering will be printed to the console. Default is TRUE.

rf.variable.importance

DEPRECATED rf.variable.importance = FALSE is no longer supported; variable importance results are always returned if the model.method is set to `pls` or `rf`.

output.summary

DEPRECATED output.summary = FALSE is no longer supported; a summary of output is always returned alongside the full performance statistics.

save.model

DEPRECATED save.model = FALSE is no longer supported; this function will always return a saved model.

vector.input

The mode of this vector of numbers will be calculated by this function

Value

list of the following:

model is a model object trained with all rows of df.

summary.model.performance is a data.frame with model performance statistics in summary format (2 rows, one with mean and one with standard deviation of all training iterations).

full.model.performance is a data.frame with model performance statistics in long format (number of rows = num.iterations)

predictions is a data.frame containing predicted values for each test set entry at each iteration of model training.

importance is a data.frame that contains variable importance for each wavelength. Only available for model.method options "rf" and "pls".

Included summary statistics:

Tuned parameters depending on the model algorithm:

Best.n.comp, the best number of components

Best.ntree, the best number of trees in an RF model

Best.mtry, the best number of variables to include at every decision point in an RF model

RMSECV, the root mean squared error of cross-validation

R2cv, the coefficient of multiple determination of cross-validation for PLSR models

RMSEP, the root mean squared error of prediction

R2p, the squared Pearson’s correlation between predicted and observed test set values

RPD, the ratio of standard deviation of observed test set values to RMSEP

RPIQ, the ratio of performance to interquartile difference

CCC, the concordance correlation coefficient

Bias, the average difference between the predicted and observed values

SEP, the standard error of prediction

R2sp, the squared Spearman’s rank correlation between predicted and observed test set values

mode of the numbers in `vector.input`

Author

Jenna Hershberger jmh579@cornell.edu

Examples

# \donttest{
library(magrittr)
ikeogu.2017 %>%
  dplyr::filter(study.name == "C16Mcal") %>%
  dplyr::rename(reference = DMC.oven,
                unique.id = sample.id) %>%
  dplyr::select(unique.id, reference, dplyr::starts_with("X")) %>%
  na.omit() %>%
  train_spectra(
    df = .,
    tune.length = 3,
    num.iterations = 3,
    best.model.metric = "RMSE",
    stratified.sampling = TRUE
  )
#> Returning model...
#> $model
#> Partial least squares regression , fitted with the kernel algorithm.
#> Call:
#> plsr(formula = reference ~ spectra, ncomp = tune.length, data = df.plsr)
#> 
#> $summary.model.performance
#>   SummaryType ModelType     RMSEp        R2p       RPD     RPIQ        CCC
#> 1        mean       pls 2.0595141 0.81708337 2.1989587 2.743205 0.88627361
#> 2          sd       pls 0.0580097 0.05160152 0.3138803 0.131658 0.04222263
#> 3        mode       pls 1.9987554 0.86292202 2.5613662 2.725104 0.92604207
#>         Bias        SEP     RMSEcv       R2cv       R2sp best.ncomp best.ntree
#> 1 -0.2452583 2.08872855 2.02653332 0.76479844 0.77019928          3         NA
#> 2  0.5196552 0.05883258 0.01782254 0.02626437 0.05009768          0         NA
#> 3 -0.2942321 2.02710800 2.00725185 0.73481010 0.82793604          3         NA
#>   best.mtry
#> 1        NA
#> 2        NA
#> 3        NA
#> 
#> $model.performance
#>   Iteration ModelType    RMSEp       R2p      RPD     RPIQ       CCC       Bias
#> 1         1       pls 1.998755 0.8629220 2.561366 2.725104 0.9260421 -0.2942321
#> 2         2       pls 2.065472 0.7611969 2.013698 2.882977 0.8419639  0.2971500
#> 3         3       pls 2.114315 0.8271312 2.021812 2.621534 0.8908148 -0.7386929
#>        SEP   RMSEcv      R2cv      R2sp best.ncomp best.ntree best.mtry
#> 1 2.027108 2.007252 0.7348101 0.8279360          3         NA        NA
#> 2 2.094771 2.029944 0.7837094 0.7382282          3         NA        NA
#> 3 2.144307 2.042404 0.7758758 0.7444336          3         NA        NA
#> 
#> $predictions
#>        Iteration ModelType unique.id     reference    predicted         
#>   [1,] "1"       "pls"     "C16Mcal_1"   "39.6210938" "37.9758040332666"
#>   [2,] "1"       "pls"     "C16Mcal_5"   "33.4427261" "30.1853801939474"
#>   [3,] "1"       "pls"     "C16Mcal_7"   "35.8100014" "37.8580854095432"
#>   [4,] "1"       "pls"     "C16Mcal_8"   "26.2774601" "29.7645595159828"
#>   [5,] "1"       "pls"     "C16Mcal_10"  "31.7993336" "33.982975941364" 
#>   [6,] "1"       "pls"     "C16Mcal_11"  "35.2299995" "37.0802591716445"
#>   [7,] "1"       "pls"     "C16Mcal_14"  "42.2379684" "40.8956958718382"
#>   [8,] "1"       "pls"     "C16Mcal_15"  "31.6499996" "31.6674988905133"
#>   [9,] "1"       "pls"     "C16Mcal_19"  "39.7091141" "38.7781676352518"
#>  [10,] "1"       "pls"     "C16Mcal_20"  "34.1300011" "32.4310520310774"
#>  [11,] "1"       "pls"     "C16Mcal_23"  "37.1399994" "36.811002746953" 
#>  [12,] "1"       "pls"     "C16Mcal_24"  "42.1911201" "39.5079662781672"
#>  [13,] "1"       "pls"     "C16Mcal_27"  "38.8980827" "39.4884913230497"
#>  [14,] "1"       "pls"     "C16Mcal_35"  "34.9000015" "38.4577176749788"
#>  [15,] "1"       "pls"     "C16Mcal_36"  "36.4031143" "36.4795725215765"
#>  [16,] "1"       "pls"     "C16Mcal_41"  "39.9312401" "39.3831259088663"
#>  [17,] "1"       "pls"     "C16Mcal_49"  "38.2799988" "37.7976125425865"
#>  [18,] "1"       "pls"     "C16Mcal_66"  "41.7697029" "38.7806505663377"
#>  [19,] "1"       "pls"     "C16Mcal_70"  "39.5100403" "38.1684591609696"
#>  [20,] "1"       "pls"     "C16Mcal_73"  "29.9218559" "31.29106282942"  
#>  [21,] "1"       "pls"     "C16Mcal_75"  "37.4979515" "40.1640232228331"
#>  [22,] "1"       "pls"     "C16Mcal_78"  "39.3699989" "38.72478249815"  
#>  [23,] "1"       "pls"     "C16Mcal_83"  "34.1562042" "33.7249425014775"
#>  [24,] "1"       "pls"     "C16Mcal_84"  "34.1899986" "34.8742953154473"
#>  [25,] "1"       "pls"     "C16Mcal_86"  "39.5099983" "41.5225544806954"
#>  [26,] "1"       "pls"     "C16Mcal_91"  "37.9599991" "37.8722240986812"
#>  [27,] "1"       "pls"     "C16Mcal_92"  "34.4248657" "35.4939093125794"
#>  [28,] "1"       "pls"     "C16Mcal_95"  "36.7374268" "36.2227820111947"
#>  [29,] "1"       "pls"     "C16Mcal_101" "41.6001511" "39.8973864705324"
#>  [30,] "1"       "pls"     "C16Mcal_104" "35.1094627" "34.35748801322"  
#>  [31,] "1"       "pls"     "C16Mcal_108" "39.1699982" "38.228157973247" 
#>  [32,] "1"       "pls"     "C16Mcal_109" "28.9400005" "31.5476900307823"
#>  [33,] "1"       "pls"     "C16Mcal_110" "23.5921307" "19.1055962717737"
#>  [34,] "1"       "pls"     "C16Mcal_115" "37.8899994" "35.8533439285928"
#>  [35,] "1"       "pls"     "C16Mcal_119" "31.1599998" "28.3684147795185"
#>  [36,] "1"       "pls"     "C16Mcal_120" "20.1389332" "16.9648863741729"
#>  [37,] "2"       "pls"     "C16Mcal_6"   "32.1251183" "34.6593389525553"
#>  [38,] "2"       "pls"     "C16Mcal_14"  "42.2379684" "40.4210943476835"
#>  [39,] "2"       "pls"     "C16Mcal_15"  "31.6499996" "32.3049885924572"
#>  [40,] "2"       "pls"     "C16Mcal_17"  "36.6281929" "38.7090748346001"
#>  [41,] "2"       "pls"     "C16Mcal_23"  "37.1399994" "37.0369473009158"
#>  [42,] "2"       "pls"     "C16Mcal_28"  "29.2099991" "31.9740579137312"
#>  [43,] "2"       "pls"     "C16Mcal_36"  "36.4031143" "36.6570651889931"
#>  [44,] "2"       "pls"     "C16Mcal_38"  "33.7883987" "36.2656045517753"
#>  [45,] "2"       "pls"     "C16Mcal_40"  "36.1300011" "37.2012360714374"
#>  [46,] "2"       "pls"     "C16Mcal_41"  "39.9312401" "39.3053483090494"
#>  [47,] "2"       "pls"     "C16Mcal_43"  "37.9103165" "39.4069428441365"
#>  [48,] "2"       "pls"     "C16Mcal_45"  "29.9400005" "33.0426849711957"
#>  [49,] "2"       "pls"     "C16Mcal_46"  "32.3829803" "33.8075241081154"
#>  [50,] "2"       "pls"     "C16Mcal_48"  "37.0624924" "37.4486927703764"
#>  [51,] "2"       "pls"     "C16Mcal_53"  "41.4599991" "38.9491617819761"
#>  [52,] "2"       "pls"     "C16Mcal_55"  "28.1499996" "32.1029799019731"
#>  [53,] "2"       "pls"     "C16Mcal_64"  "27.2600002" "28.6073894488928"
#>  [54,] "2"       "pls"     "C16Mcal_65"  "38.8069687" "37.3767781922965"
#>  [55,] "2"       "pls"     "C16Mcal_66"  "41.7697029" "38.6039193923932"
#>  [56,] "2"       "pls"     "C16Mcal_70"  "39.5100403" "38.1115760248322"
#>  [57,] "2"       "pls"     "C16Mcal_73"  "29.9218559" "31.9196636028002"
#>  [58,] "2"       "pls"     "C16Mcal_77"  "33.8568802" "30.3830486864091"
#>  [59,] "2"       "pls"     "C16Mcal_79"  "35.9900017" "39.2652727319261"
#>  [60,] "2"       "pls"     "C16Mcal_84"  "34.1899986" "35.0214107604401"
#>  [61,] "2"       "pls"     "C16Mcal_86"  "39.5099983" "41.2034677763275"
#>  [62,] "2"       "pls"     "C16Mcal_87"  "35.0520935" "37.9401200207589"
#>  [63,] "2"       "pls"     "C16Mcal_91"  "37.9599991" "37.7958095718426"
#>  [64,] "2"       "pls"     "C16Mcal_95"  "36.7374268" "36.330791316579" 
#>  [65,] "2"       "pls"     "C16Mcal_97"  "34.2159386" "35.5263709601503"
#>  [66,] "2"       "pls"     "C16Mcal_100" "34.8311043" "36.5400097908371"
#>  [67,] "2"       "pls"     "C16Mcal_101" "41.6001511" "39.5349464108344"
#>  [68,] "2"       "pls"     "C16Mcal_103" "39.3523369" "39.0495506639272"
#>  [69,] "2"       "pls"     "C16Mcal_105" "41.1385422" "37.8870506449326"
#>  [70,] "2"       "pls"     "C16Mcal_106" "40.3611107" "37.431149223532" 
#>  [71,] "2"       "pls"     "C16Mcal_108" "39.1699982" "38.2773500596091"
#>  [72,] "2"       "pls"     "C16Mcal_119" "31.1599998" "29.1429512027063"
#>  [73,] "3"       "pls"     "C16Mcal_11"  "35.2299995" "36.9857739690787"
#>  [74,] "3"       "pls"     "C16Mcal_12"  "41.9791336" "41.1573530927889"
#>  [75,] "3"       "pls"     "C16Mcal_15"  "31.6499996" "31.6595806061607"
#>  [76,] "3"       "pls"     "C16Mcal_16"  "36.3796349" "38.5317052280193"
#>  [77,] "3"       "pls"     "C16Mcal_18"  "39.5632248" "37.1025425768922"
#>  [78,] "3"       "pls"     "C16Mcal_19"  "39.7091141" "38.6194125828423"
#>  [79,] "3"       "pls"     "C16Mcal_23"  "37.1399994" "36.855058067351" 
#>  [80,] "3"       "pls"     "C16Mcal_31"  "32.4087524" "33.3109794668438"
#>  [81,] "3"       "pls"     "C16Mcal_36"  "36.4031143" "36.5644040837359"
#>  [82,] "3"       "pls"     "C16Mcal_40"  "36.1300011" "37.0876089919321"
#>  [83,] "3"       "pls"     "C16Mcal_43"  "37.9103165" "39.4741666283103"
#>  [84,] "3"       "pls"     "C16Mcal_45"  "29.9400005" "32.3359748539552"
#>  [85,] "3"       "pls"     "C16Mcal_50"  "37.7099991" "35.6198575435607"
#>  [86,] "3"       "pls"     "C16Mcal_51"  "39.348053"  "37.3069991545157"
#>  [87,] "3"       "pls"     "C16Mcal_56"  "38.4799995" "37.838246782011" 
#>  [88,] "3"       "pls"     "C16Mcal_57"  "40.5781174" "38.6214325770801"
#>  [89,] "3"       "pls"     "C16Mcal_61"  "39.4031334" "40.7235829690078"
#>  [90,] "3"       "pls"     "C16Mcal_62"  "29.073864"  "25.2664481873008"
#>  [91,] "3"       "pls"     "C16Mcal_66"  "41.7697029" "38.6373217808577"
#>  [92,] "3"       "pls"     "C16Mcal_69"  "31.9607944" "29.7826670031293"
#>  [93,] "3"       "pls"     "C16Mcal_71"  "40.0252266" "38.560264130916" 
#>  [94,] "3"       "pls"     "C16Mcal_79"  "35.9900017" "39.321665481591" 
#>  [95,] "3"       "pls"     "C16Mcal_82"  "34.8699989" "29.9492858456461"
#>  [96,] "3"       "pls"     "C16Mcal_83"  "34.1562042" "33.3810179719606"
#>  [97,] "3"       "pls"     "C16Mcal_85"  "33.1330147" "32.8880378719808"
#>  [98,] "3"       "pls"     "C16Mcal_92"  "34.4248657" "35.4411613256911"
#>  [99,] "3"       "pls"     "C16Mcal_93"  "34.8271332" "35.8972092234424"
#> [100,] "3"       "pls"     "C16Mcal_98"  "38.8300018" "40.1326841393942"
#> [101,] "3"       "pls"     "C16Mcal_102" "31.114912"  "30.8367837550963"
#> [102,] "3"       "pls"     "C16Mcal_106" "40.3611107" "37.4934809389549"
#> [103,] "3"       "pls"     "C16Mcal_107" "38.8699989" "37.4885166638299"
#> [104,] "3"       "pls"     "C16Mcal_108" "39.1699982" "38.2319952801923"
#> [105,] "3"       "pls"     "C16Mcal_110" "23.5921307" "18.6506372137214"
#> [106,] "3"       "pls"     "C16Mcal_112" "40.9370956" "39.250102647691" 
#> [107,] "3"       "pls"     "C16Mcal_116" "43.2331314" "41.7116248159891"
#> [108,] "3"       "pls"     "C16Mcal_119" "31.1599998" "28.1532490825658"
#> 
#> $importance
#>     wavelength Iteration ModelType    Overall
#> 1         X350         1       pls 0.03065491
#> 2         X351         1       pls 0.02941634
#> 3         X352         1       pls 0.03440090
#> 4         X353         1       pls 0.03310131
#> 5         X354         1       pls 0.03261624
#> 6         X355         1       pls 0.03476287
#> 7         X356         1       pls 0.03492789
#> 8         X357         1       pls 0.03637971
#> 9         X358         1       pls 0.03607341
#> 10        X359         1       pls 0.03613006
#> 11        X360         1       pls 0.03916783
#> 12        X361         1       pls 0.04107061
#> 13        X362         1       pls 0.04187890
#> 14        X363         1       pls 0.04160233
#> 15        X364         1       pls 0.04166290
#> 16        X365         1       pls 0.04178564
#> 17        X366         1       pls 0.04095651
#> 18        X367         1       pls 0.04206731
#> 19        X368         1       pls 0.04417454
#> 20        X369         1       pls 0.04531990
#> 21        X370         1       pls 0.04517321
#> 22        X371         1       pls 0.04583639
#> 23        X372         1       pls 0.04686563
#> 24        X373         1       pls 0.04738602
#> 25        X374         1       pls 0.04705522
#> 26        X375         1       pls 0.04946838
#> 27        X376         1       pls 0.05008552
#> 28        X377         1       pls 0.05105102
#> 29        X378         1       pls 0.04969948
#> 30        X379         1       pls 0.04834912
#> 31        X380         1       pls 0.05188571
#> 32        X381         1       pls 0.05216037
#> 33        X382         1       pls 0.05119620
#> 34        X383         1       pls 0.05337564
#> 35        X384         1       pls 0.05465650
#> 36        X385         1       pls 0.05363718
#> 37        X386         1       pls 0.05258051
#> 38        X387         1       pls 0.05523107
#> 39        X388         1       pls 0.05410153
#> 40        X389         1       pls 0.05410779
#> 41        X390         1       pls 0.05465887
#> 42        X391         1       pls 0.05522291
#> 43        X392         1       pls 0.05678501
#> 44        X393         1       pls 0.05755872
#> 45        X394         1       pls 0.05721063
#> 46        X395         1       pls 0.05784821
#> 47        X396         1       pls 0.05922115
#> 48        X397         1       pls 0.05989602
#> 49        X398         1       pls 0.05963396
#> 50        X399         1       pls 0.06079326
#> 51        X400         1       pls 0.06236197
#> 52        X401         1       pls 0.06340402
#> 53        X402         1       pls 0.06461249
#> 54        X403         1       pls 0.06430192
#> 55        X404         1       pls 0.06513903
#> 56        X405         1       pls 0.06620422
#> 57        X406         1       pls 0.06752711
#> 58        X407         1       pls 0.06768090
#> 59        X408         1       pls 0.06878692
#> 60        X409         1       pls 0.06856927
#> 61        X410         1       pls 0.06890526
#> 62        X411         1       pls 0.06906749
#> 63        X412         1       pls 0.06943575
#> 64        X413         1       pls 0.06873410
#> 65        X414         1       pls 0.06945458
#> 66        X415         1       pls 0.07063529
#> 67        X416         1       pls 0.07072705
#> 68        X417         1       pls 0.07043618
#> 69        X418         1       pls 0.07119238
#> 70        X419         1       pls 0.07191677
#> 71        X420         1       pls 0.07205225
#> 72        X421         1       pls 0.07235481
#> 73        X422         1       pls 0.07289053
#> 74        X423         1       pls 0.07381071
#> 75        X424         1       pls 0.07477647
#> 76        X425         1       pls 0.07617255
#> 77        X426         1       pls 0.07676705
#> 78        X427         1       pls 0.07746323
#> 79        X428         1       pls 0.07871132
#> 80        X429         1       pls 0.07972230
#> 81        X430         1       pls 0.08092815
#> 82        X431         1       pls 0.08167483
#> 83        X432         1       pls 0.08252184
#> 84        X433         1       pls 0.08354534
#> 85        X434         1       pls 0.08416049
#> 86        X435         1       pls 0.08496768
#> 87        X436         1       pls 0.08536459
#> 88        X437         1       pls 0.08547003
#> 89        X438         1       pls 0.08575684
#> 90        X439         1       pls 0.08610241
#> 91        X440         1       pls 0.08609144
#> 92        X441         1       pls 0.08617123
#> 93        X442         1       pls 0.08606270
#> 94        X443         1       pls 0.08615394
#> 95        X444         1       pls 0.08624085
#> 96        X445         1       pls 0.08608917
#> 97        X446         1       pls 0.08583322
#> 98        X447         1       pls 0.08607675
#> 99        X448         1       pls 0.08597194
#> 100       X449         1       pls 0.08610824
#> 101       X450         1       pls 0.08641160
#> 102       X451         1       pls 0.08640764
#> 103       X452         1       pls 0.08663500
#> 104       X453         1       pls 0.08686543
#> 105       X454         1       pls 0.08702454
#> 106       X455         1       pls 0.08731024
#> 107       X456         1       pls 0.08758371
#> 108       X457         1       pls 0.08772004
#> 109       X458         1       pls 0.08793538
#> 110       X459         1       pls 0.08800683
#> 111       X460         1       pls 0.08819525
#> 112       X461         1       pls 0.08831824
#> 113       X462         1       pls 0.08828073
#> 114       X463         1       pls 0.08827365
#> 115       X464         1       pls 0.08819993
#> 116       X465         1       pls 0.08804116
#> 117       X466         1       pls 0.08782952
#> 118       X467         1       pls 0.08750955
#> 119       X468         1       pls 0.08734643
#> 120       X469         1       pls 0.08727688
#> 121       X470         1       pls 0.08678730
#> 122       X471         1       pls 0.08633365
#> 123       X472         1       pls 0.08606814
#> 124       X473         1       pls 0.08567050
#> 125       X474         1       pls 0.08526253
#> 126       X475         1       pls 0.08505495
#> 127       X476         1       pls 0.08479786
#> 128       X477         1       pls 0.08466648
#> 129       X478         1       pls 0.08460856
#> 130       X479         1       pls 0.08450183
#> 131       X480         1       pls 0.08430539
#> 132       X481         1       pls 0.08419809
#> 133       X482         1       pls 0.08417656
#> 134       X483         1       pls 0.08418090
#> 135       X484         1       pls 0.08421475
#> 136       X485         1       pls 0.08430804
#> 137       X486         1       pls 0.08430797
#> 138       X487         1       pls 0.08442458
#> 139       X488         1       pls 0.08427991
#> 140       X489         1       pls 0.08418280
#> 141       X490         1       pls 0.08412807
#> 142       X491         1       pls 0.08398496
#> 143       X492         1       pls 0.08366754
#> 144       X493         1       pls 0.08341705
#> 145       X494         1       pls 0.08319492
#> 146       X495         1       pls 0.08292474
#> 147       X496         1       pls 0.08258787
#> 148       X497         1       pls 0.08213628
#> 149       X498         1       pls 0.08161968
#> 150       X499         1       pls 0.08113378
#> 151       X500         1       pls 0.08062201
#> 152       X501         1       pls 0.08003950
#> 153       X502         1       pls 0.07942165
#> 154       X503         1       pls 0.07879370
#> 155       X504         1       pls 0.07818462
#> 156       X505         1       pls 0.07743613
#> 157       X506         1       pls 0.07692178
#> 158       X507         1       pls 0.07628174
#> 159       X508         1       pls 0.07543950
#> 160       X509         1       pls 0.07472929
#> 161       X510         1       pls 0.07410618
#> 162       X511         1       pls 0.07341634
#> 163       X512         1       pls 0.07272044
#> 164       X513         1       pls 0.07205159
#> 165       X514         1       pls 0.07144360
#> 166       X515         1       pls 0.07072197
#> 167       X516         1       pls 0.07005738
#> 168       X517         1       pls 0.06933525
#> 169       X518         1       pls 0.06862030
#> 170       X519         1       pls 0.06792409
#> 171       X520         1       pls 0.06720595
#> 172       X521         1       pls 0.06642453
#> 173       X522         1       pls 0.06568059
#> 174       X523         1       pls 0.06494857
#> 175       X524         1       pls 0.06413616
#> 176       X525         1       pls 0.06338598
#> 177       X526         1       pls 0.06256557
#> 178       X527         1       pls 0.06169854
#> 179       X528         1       pls 0.06083551
#> 180       X529         1       pls 0.06003039
#> 181       X530         1       pls 0.05913191
#> 182       X531         1       pls 0.05819154
#> 183       X532         1       pls 0.05728573
#> 184       X533         1       pls 0.05635742
#> 185       X534         1       pls 0.05538626
#> 186       X535         1       pls 0.05434409
#> 187       X536         1       pls 0.05327600
#> 188       X537         1       pls 0.05213039
#> 189       X538         1       pls 0.05097391
#> 190       X539         1       pls 0.04976476
#> 191       X540         1       pls 0.04848917
#> 192       X541         1       pls 0.04718449
#> 193       X542         1       pls 0.04587320
#> 194       X543         1       pls 0.04449677
#> 195       X544         1       pls 0.04324728
#> 196       X545         1       pls 0.04197458
#> 197       X546         1       pls 0.04068299
#> 198       X547         1       pls 0.03934143
#> 199       X548         1       pls 0.03801278
#> 200       X549         1       pls 0.03667381
#> 201       X550         1       pls 0.03535376
#> 202       X551         1       pls 0.03404872
#> 203       X552         1       pls 0.03276303
#> 204       X553         1       pls 0.03153625
#> 205       X554         1       pls 0.03034151
#> 206       X555         1       pls 0.02918747
#> 207       X556         1       pls 0.02807341
#> 208       X557         1       pls 0.02698468
#> 209       X558         1       pls 0.02591947
#> 210       X559         1       pls 0.02491936
#> 211       X560         1       pls 0.02401029
#> 212       X561         1       pls 0.02315709
#> 213       X562         1       pls 0.02233908
#> 214       X563         1       pls 0.02158419
#> 215       X564         1       pls 0.02093392
#> 216       X565         1       pls 0.02030899
#> 217       X566         1       pls 0.01971703
#> 218       X567         1       pls 0.01916724
#> 219       X568         1       pls 0.01866955
#> 220       X569         1       pls 0.01819969
#> 221       X570         1       pls 0.01774829
#> 222       X571         1       pls 0.01734926
#> 223       X572         1       pls 0.01695315
#> 224       X573         1       pls 0.01659090
#> 225       X574         1       pls 0.01626621
#> 226       X575         1       pls 0.01594766
#> 227       X576         1       pls 0.01565753
#> 228       X577         1       pls 0.01537906
#> 229       X578         1       pls 0.01511687
#> 230       X579         1       pls 0.01487086
#> 231       X580         1       pls 0.01464582
#> 232       X581         1       pls 0.01440942
#> 233       X582         1       pls 0.01419936
#> 234       X583         1       pls 0.01398947
#> 235       X584         1       pls 0.01379112
#> 236       X585         1       pls 0.01361325
#> 237       X586         1       pls 0.01344748
#> 238       X587         1       pls 0.01326934
#> 239       X588         1       pls 0.01310923
#> 240       X589         1       pls 0.01297067
#> 241       X590         1       pls 0.01281711
#> 242       X591         1       pls 0.01266377
#> 243       X592         1       pls 0.01252424
#> 244       X593         1       pls 0.01237988
#> 245       X594         1       pls 0.01225212
#> 246       X595         1       pls 0.01214064
#> 247       X596         1       pls 0.01201525
#> 248       X597         1       pls 0.01190647
#> 249       X598         1       pls 0.01180099
#> 250       X599         1       pls 0.01168906
#>  [ reached 'max' / getOption("max.print") -- omitted 6203 rows ]
#> 
# }